Split management module
