+++ lib/utilities/reportHandler.js
@@ -131,6 +131,7 @@ function reportHandler(clientIP, req, res, log) {
         return;
     }
 
+    // TODO propagate value of req.headers['x-scal-report-skip-cache']
     async.parallel({
         getUUID: cb => metadata.getUUID(log, cb),
         getMDDiskUsage: cb => metadata.getDiskUsage(log, cb),
